requirements.txt
# Core LangGraph & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-vertexai>=2.0.0
langchain-google-community>=2.0.0

# Google Cloud
google-cloud-bigquery>=3.25.0
google-cloud-firestore>=2.19.0
google-cloud-aiplatform>=1.70.0
google-auth>=2.35.0

# Vertex AI
vertexai>=1.70.0

# Web & API
requests>=2.32.0
aiohttp>=3.10.0

# Data Processing
pandas>=2.2.0
pydantic>=2.9.0

# Frontend
streamlit>=1.40.0

# Utilities
python-dotenv>=1.0.1
tenacity>=9.0.0

# Development
pytest>=8.3.0
black>=24.10.0
